Send a request to an API

This function implements an opinionated framework for making API calls. It is intended to be used inside an API client package. It serves as a wrapper around the req_ family of functions, such as httr2::request().

Usage

call_api(
  base_url,
  endpoint,
  query = NULL,
  body = NULL,
  mime_type = NULL,
  method = NULL,
  api_case = c("snake_case", "camelCase", "UpperCamel", "SCREAMING_SNAKE", "alllower",
    "ALLUPPER", "lowerUPPER", "UPPERlower", "Sentence case", "Title Case"),
  authenticator = NULL,
  authenticator_args = list(),
  response_parser = NULL,
  response_parser_args = list(),
  error_handler = NULL,
  error_handler_args = list(),
  user_agent = "nectar (https://jonthegeek.github.io/nectar/)"
)

Arguments

base_url

The part of the url that is shared by all calls to the API. In some cases there may be a family of base URLs, from which you will need to choose one.

endpoint

The path to an API endpoint. Optionally, a list with the path plus variables to glue::glue() into the path.

query

An optional list of parameters to pass in the query portion of the request.

body

An object to use as the body of the request. If any component of the body is a path, pass it through fs::path() or otherwise give it the class "fs_path" to indicate that it is a path.

mime_type

A character scalar indicating the mime type of any files present in the body. Some APIs allow you to leave this as NULL for them to guess.

method

If the method is something other than GET or POST, supply it. Case is ignored.

api_case

The case standard used by this API. The possible values are self-descriptive. Defaults to "snake_case".

authenticator

An authenticator function. More details to come in a future update to this package.

authenticator_args

Arguments for the authenticator function.

response_parser

A response parser function. More details to come in a future update to this package.

response_parser_args

Arguments for the response parser function.

error_handler

An error handler function. More details to come in a future update to this package.

error_handler_args

Arguments for the error handler function.

user_agent

A string to identify where this request is coming from. It's polite to set the user agent to identify your package, such as "MyPackage (https://mypackage.com)".

Value

The response from the API, parsed by the response_parser.