Send data in request body

Automatically choose between httr2::req_body_json() and httr2::req_body_multipart() based on the content of the body. This is currently experimental and needs to be tested on more APIs.

Usage

.req_body_auto(req, body, mime_type = NULL)

Arguments

req

(httr2_request) A httr2::request() object.

body

(multiple types) An object to use as the body of the request. If any component of the body is a path, pass it through fs::path() or otherwise give it the class "fs_path" to indicate that it is a path.

mime_type

(length-1 character) The mime type of any files present in the body. Some APIs allow you to leave this as NULL for them to guess.

Value

A modified HTTP request.