Parameters used in multiple functions

Reused parameter definitions are gathered here for easier editing.

Arguments

base_url

The part of the url that is shared by all calls to the API. In some cases there may be a family of base URLs, from which you will need to choose one.

body

An object to use as the body of the request. If any component of the body is a path, pass it through fs::path() or otherwise give it the class "fs_path" to indicate that it is a path.

case

The case standard to apply. The possible values are self-descriptive. Defaults to "snake_case".

depth

The current recursion depth.

method

If the method is something other than GET or POST, supply it. Case is ignored.

mime_type

A character scalar indicating the mime type of any files present in the body. Some APIs allow you to leave this as NULL for them to guess.

path

The route to an API endpoint. Optionally, a list with the path plus variables to glue::glue() into the path.

query

An optional list of parameters to pass in the query portion of the request.

req

An httr2::request() object.

response_parser

A function to parse the server response (resp). Defaults to httr2::resp_body_json(), since JSON responses are common. Set this to NULL to return the raw response from httr2::req_perform().

response_parser_args

An optional list of arguments to pass to the response_parser function (in addition to resp).

security_fn

A function to use to authenticate the request. By default (NULL), no authentication is performed.

security_args

An optional list of arguments to the security_fn function.

user_agent

A string to identify where this request is coming from. It's polite to set the user agent to identify your package, such as "MyPackage (https://mypackage.com)".

x

The object to update.