Setup a basic API request

For a given API, the base_url and user agent will generally be the same for every call to that API. Use this function to prepare that piece of the request once for easy reuse.

Usage

req_init(base_url, ..., additional_user_agent = NULL)

Arguments

base_url

(length-1 character) The part of the url that is shared by all calls to the API. In some cases there may be a family of base URLs, from which you will need to choose one.

...

These dots are for future extensions and must be empty.

additional_user_agent

(length-1 character) A string to identify where a request is coming from. We automatically include information about your package and nectar, but use this to provide additional details. Default NULL.

Value

A httr2::request() object.

Examples

req_init("https://example.com")
#> <httr2_request>
#> GET https://example.com
#> Body: empty
#> Options:
#> • useragent: 'httr2/1.0.7 r-curl/6.0.1 libcurl/8.5.0 nectar/0.0.0.9003
#> (https://nectar.api2r.org)'
req_init(
  "https://example.com",
  additional_user_agent = "my_api_client (https://my.api.client)"
)
#> <httr2_request>
#> GET https://example.com
#> Body: empty
#> Options:
#> • useragent: 'my_api_client (https://my.api.client) nectar/0.0.0.9003
#> (https://nectar.api2r.org)'